+++ src/utils.js
@@ -1,8 +1,3 @@
-const RenderPosition = {
-  AFTERBEGIN: `afterbegin`,
-  BEFOREEND: `beforeend`
-};
-
 const formatTime = (date) => {
   const hours = String(date.getHours()).padStart(2, `0`);
   const minutes = String(date.getMinutes()).padStart(2, `0`);
@@ -10,22 +5,4 @@ const formatTime = (date) => {
   return `${hours}:${minutes}`;
 };
 
-const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-
-  return newElement.firstChild;
-};
-
-const render = (container, element, place) => {
-  switch (place) {
-    case RenderPosition.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case RenderPosition.BEFOREEND:
-      container.append(element);
-      break;
-  }
-};
-
-export {RenderPosition, formatTime, createElement, render};
+export {formatTime};
